add file for Board game class
